continue changing to unsigned int for color
it started with colors not working properly cause bit shifting examples depend on habving an unsigned int for colors. so now iam redoing the whole reader to split iinto functions and changing the int array to an array for z values and color in a spearate unsigned int value. worked on fixing the get next line to return -1 on error and refactoring the functions in the reader. the basic version still works here